third_party/WebKit/LayoutTests/fast/events/dispatch-synthetic-keyboardevent-no-action.html

<!DOCTYPE html>
<script src="../../resources/js-test.js"></script>
<p id="description"></p>
<form>
<input type="radio" id="radio1" name="foo" value="A" checked/>
<input type="radio" id="radio2" name="foo" value="B" />
</form>
<pre id="console"></pre>
<script>
description("Tests to ensure that default action does not occur for untrusted events.");

var node = document.getElementById('radio1');
node.getBoundingClientRect();

var event = new KeyboardEvent('keydown', { key: 'ArrowDown' } );
node.dispatchEvent(event);
shouldBeTrue("document.getElementById('radio1').checked");
shouldBeFalse("document.getElementById('radio2').checked");
</script>
